✨ Feat : event part service.py로 notification.py와 crud.py를 합쳤습니다.
